Move NonZeroExitCodeError to common
